+++ src/webserver/database/sqlpage_functions/mod.rs
@@ -7,6 +7,7 @@ use sqlparser::ast::FunctionArg;
 
 use crate::webserver::http_request_info::RequestInfo;
 
+use super::sql::function_args_to_stmt_params;
 use super::syntax_tree::SqlPageFunctionCall;
 use super::syntax_tree::StmtParam;
 
@@ -25,4 +26,9 @@ pub(super) fn func_call_to_param(func_name: &str, arguments: &mut [FunctionArg])
             |e| StmtParam::Error(format!("{e:#}")),
             StmtParam::FunctionCall,
         )
+}
+
+pub(super) fn are_params_extractable(arguments: &[FunctionArg]) -> bool {
+    let mut mutable_copy = arguments.to_vec();
+    function_args_to_stmt_params(&mut mutable_copy).is_ok()
 }