+++ tests/index.rs
@@ -165,10 +165,10 @@ async fn test_files() {
             );
         } else if stem.starts_with("error_") {
             let rest = stem.strip_prefix("error_").unwrap();
-            let expected_str = rest.replace('_', " ");
+            let expected_str = rest.replace('_', " ").to_lowercase();
             assert!(
                 lowercase_body.contains(&expected_str),
-                "{req_str}\n{body}\nexpected to contain: {expected_str}"
+                "{req_str}\n{lowercase_body}\nexpected to contain: {expected_str}"
             );
             assert!(
                 lowercase_body.contains("error"),
@@ -520,7 +520,7 @@ pub fn test_config() -> AppConfig {
 
 fn init_log() {
     let _ = env_logger::builder()
+        .parse_default_env()
         .is_test(true)
-        .filter(Some("sqlpage"), log::LevelFilter::Trace)
         .try_init();
 }